# internal/httpapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AIV2API",
    "version": "1.5.0",
    "description": "Image, video, and audio generation API. Video, audio, and asynchronous image endpoints return tasks.",
    "license": {"name": "All rights reserved", "identifier": "LicenseRef-Proprietary"}
  },
  "servers": [{"url": "/", "description": "Current host"}],
...
  "paths": {
...
    "/v1/models": {
      "get": {"tags": ["Models"], "operationId": "listModels", "summary": "List models allowed to this API key", "responses": {"200": {"description": "Curated public models allowed to this API key", "content": {"application/json": {"schema": {"type": "object", "properties": {"object": {"const": "list"}, "data": {"type": "array", "items": {"$ref": "#/components/schemas/Model"}}}}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}}}
    },
...
  },
  "components": {
    "securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "API key"}},
...
    "schemas": {
...
      "Model": {"type": "object", "required": ["id", "object", "created", "owned_by"], "additionalProperties": false, "properties": {"id": {"type": "string"}, "object": {"const": "model"}, "created": {"type": "integer", "format": "int64"}, "owned_by": {"const": "aiv2api"}}},
...
      "Error": {"type": "object", "required": ["message", "type", "code"], "properties": {"message": {"type": "string"}, "type": {"type": "string"}, "code": {"type": "string"}}},
      "ErrorResponse": {"type": "object", "required": ["error"], "properties": {"error": {"$ref": "#/components/schemas/Error"}}}
    },
    "responses": {
...
      "Unauthorized": {"description": "Missing or invalid Bearer API key", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}, "example": {"error": {"message": "missing Bearer API key", "type": "invalid_api_key", "code": "invalid_api_key"}}}}},
...
    }
  },
  "security": [{"bearerAuth": []}]
}